Work with absolute paths
We assume that the template and output directory is always relative to the
script location and not where you call it, this makes it easier to set up an
alias. If this is wrong and you want relative paths from current directory,
simply remove `File.dirname(__FILE__)` below and when generating.